move things away from TrackProducer
